refactor: 拆分 enums.py 为领域子模块包
ginkgo/enums.py (917 行 / 52 枚举 + EnumBase 基类) 按事件链路与既有
分节注释拆分为 10 个领域子模块：

  base       EnumBase 基类（转换方法）
  trading    交易订单（方向/订单/状态/资金/币种）
  data       数据源（价格/源/频率/市场/复权）
  event      事件链路
  strategy   策略（态度/类型/模型/分析器集合）
  execution  引擎与执行（架构/模式/账户/状态）
  portfolio  Portfolio 运行时（Phase 5 优雅重启）
  system     系统通用（文件/组件/日志/环境/权限/部署）
  account    实盘账户（交易所/状态/订阅/Broker 状态机）
  user       用户与通知（US2）

- __init__.py 用 `from .X import *` 聚合 re-export 全部 53 个符号，
  928 处 `from ginkgo.enums import XXX` 旧路径零感知（向后兼容契约）。
- 各子模块按实际引用生成精确 import + __all__（str 子类不误 import EnumBase）。
- 顺带丢弃死 import IntEnum（原文件仅 import 行/注释，无 class 使用）。
- 新增 test_enums_registry.py：53 符号顶层可导入 + 关键值/双L别名/
  classmethod/str 身份/命名空间类/领域路径 identity 共 8 个契约测试。

行为零变更验证：
  tracer bullet 8 + 既有枚举测试 12 全过；
  py_compile 814 文件通过；
  启动路径 smoke (user_crud/containers/user_cli) 36 passed。
